+++ CHANGELOG.md
@@ -87,3 +87,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 #### Added
 - Delete wishlist
 - Add home partner app
+- Delete product